Refactor service imports and add LLM and RAG services with corresponding tests
