+++ tapio/models/__init__.py
@@ -1,6 +1,4 @@
 # This file initializes the models module.
 from tapio.models.document import Document
-from tapio.models.llm_service import LLMService
-from tapio.models.rag_service import RAGService
 
-__all__ = ["Document", "LLMService", "RAGService"]
+__all__ = ["Document"]
